requirements.txt
androidviewclient
python-dotenv
openai
hangul-utils